Move recipe components to dedicated package
